LayoutTests/inspector/inspector-backend-commands-generation-expected.txt

Tests correctness of code generated by InspectorBackend.

InspectorBackend.registerEnum("Page.ResourceType", {Document: "Document", XHR: "XHR", ConsoleAPI: "console-api", Other: "other"});
InspectorBackend.registerCommand("Page.addScriptToEvaluateOnLoad", [{"name": "scriptSource", "type": "string", "optional": false}], ["identifier"]);
InspectorBackend.registerEvent("Page.domContentEventFired", ["timestamp"]);
InspectorBackend.registerPageDispatcher = InspectorBackend.registerDomainDispatcher.bind(InspectorBackend, "Page");
